+++ WebPlaylistToSpotify/Program.cs
@@ -8,12 +8,8 @@ using WebPlaylistToSpotify.Model;
 try
 {
     var appConfig = new AppConfig();
-
-    var browser = new SystemBrowser(appConfig.SpotifyClientId);
-    var queryString = await browser.InvokeAsync();
-    var parsedQuery = HttpUtility.ParseQueryString(queryString);
-    var token = await browser.GetToken(parsedQuery["code"]);
-
+    
+    string token = await GetAuthToken(appConfig);
     var spotify = new SpotifyClient(token);
 
     Console.WriteLine("Starting...");
@@ -88,4 +84,21 @@ static async Task<FullPlaylist> CreatePlaylist(AppConfig appConfig, SpotifyClien
 
     Console.WriteLine($"Created spotify playlist: {newPlaylistName}");
     return playlist;
+}
+
+static async Task<string> GetAuthToken(AppConfig appConfig)
+{
+    var browser = new SpotifyOAuthBrowser(appConfig.SpotifyClientId);
+    var resultQueryString = await browser.StartFlow();
+    var parsedQuery = HttpUtility.ParseQueryString(resultQueryString);
+
+    var code = parsedQuery["code"];
+
+    if (code == null)
+    {
+        throw new InvalidOperationException("Missing authorisation code");
+    }
+
+    var token = await browser.GetToken(code);
+    return token;
 }
